reducing memory usage (int->short)

+++ minimax/TwoPlayerGame.java
@@ -2,7 +2,7 @@ package minimax;
 
 public interface TwoPlayerGame {
 
-    public int point(Table t, int player);
-    public Table turn(Table t, int player, Move move);
-    public Move[] possibleMoves(Table t, int player);
+    public int point(Table t, short player);
+    public Table turn(Table t, short player, Move move);
+    public Move[] possibleMoves(Table t, short player);
 }
